+++ infra/resources/ced/_modules/dns/variables.tf
@@ -16,8 +16,50 @@ variable "virtual_network" {
   description = "Virtual network where to attach private dns zones"
 }
 
-variable "override_link_names" {
-  type        = map(string)
-  description = "Override link names"
-  default     = {}
+variable "dkim1_record" {
+  type        = string
+  description = "DKIM1 record value"
+  sensitive   = true
+}
+
+variable "dkim2_record" {
+  type        = string
+  description = "DKIM2 record value"
+  sensitive   = true
+}
+
+variable "dkim3_record" {
+  type        = string
+  description = "DKIM3 record value"
+  sensitive   = true
+}
+
+variable "cmf_mx_1_record" {
+  type        = string
+  description = "Custom Mail From MX 1 record value"
+  sensitive   = true
+}
+
+variable "cmf_txt_1_record" {
+  type        = string
+  description = "Custom Mail From TXT 1 record value"
+  sensitive   = true
+}
+
+variable "cmf_mx_2_record" {
+  type        = string
+  description = "Custom Mail From MX 2 record value"
+  sensitive   = true
+}
+
+variable "cmf_txt_2_record" {
+  type        = string
+  description = "Custom Mail From TXT 2 record value"
+  sensitive   = true
+}
+
+variable "dmarc_record" {
+  type        = string
+  description = "DMARC record value"
+  sensitive   = true
 }